Move Display implementation, add utf8_len method
