+++ src/attribute.rs
@@ -1,3 +1,4 @@
+use std::fmt;
 use std::borrow::Cow;
 
 static EMPTY_ATTRS: &'static [Attribute<'static>] = &[];
@@ -11,5 +12,18 @@ pub struct Attribute<'a> {
 impl<'a> Attribute<'a> {
     pub fn none() -> Cow<'a, [Attribute<'a>]> {
         Cow::Borrowed(EMPTY_ATTRS)
+    }
+}
+
+impl<'a> fmt::Display for Attribute<'a> {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        // TODO: Escaping
+        write!(f, "{}=\"{}\"", self.name, self.value)
+    }
+}
+
+impl<'a> Attribute<'a> {
+    pub fn utf8_len(&self) -> usize {
+        self.name.len() + self.value.len() + 3
     }
 }
