+++ src/lib.rs
@@ -3,7 +3,6 @@ pub mod macros;
 mod event;
 mod attribute;
 mod write;
-mod fmt;
 
 pub use event::Event;
 pub use attribute::Attribute;
